Initial work.
This is all the work I did up to the point where I started working on the actual project.  At first I thought we'd be required to transfer files over the network, but that is not a requirement.  This code includes the initial client and server, and it transfers ASCII (maybe binary too) files less than 1024 characters.  After that, which is the message size, a read from the network reads in more than one message, and the program can't interpret that properly yet.  I could implement that by putting incoming messages into a circular buffer, then pulling out the beginning of the message, reading that message size, and removing that single message from the buffer.